day4-7/pyproject.toml
[project]
name = "blog-api"
version = "0.1.0"
description = "Blog API — FastAPI training Day 4-7"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # Pydantic & settings  (Day 4)
    "pydantic[email]>=2.7.0",
    "pydantic-settings>=2.3.0",
    # Database & migrations  (Day 5)
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    # Security & Auth  (Day 6)
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    # Logging
    "structlog>=24.2.0",
    "bcrypt==4.0.1",
]

# Đã sửa lại cú pháp chuẩn của PEP 621
[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0"
]

# BÙA HỘ MỆNH CHO UV: Báo cho hệ thống biết đây là App, không phải Thư viện (Package)
[tool.uv]
package = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
